Delete and Escape Binding Fix Validation again

+++ CompuStore.Clients/ViewModels/ClientPaymentMainViewModel.cs
@@ -83,6 +83,8 @@ namespace CompuStore.Clients.ViewModels
         }
         private void Delete()
         {
+            if (SelectedItem == null)
+                return;
             if (!Messages.Delete("فاتورة ايصال نقدية رقم " + SelectedItem.Number.ToString())) return;
             if(!_clientPaymentService.Delete(SelectedItem))
             {
